made server good looking good cooking, added new config parser, needs better static and indexing handlers thoooooo

+++ inc/Config/Config.hpp
@@ -29,7 +29,7 @@ struct ServerConfig
     int port;
     std::string host;
     std::string server_name;
-    int client_max_body_size;
+    size_t client_max_body_size;
     std::map<int, std::string> error_pages;
     std::vector<LocConfig> locations;
 };
